Split GFQL temporal text helpers
