+++ graphistry/compute/gfql/temporal/__init__.py
@@ -1,0 +1,1 @@
+"""Temporal text parsing and folding helpers for GFQL."""
