+++ graphistry/compute/gfql/temporal_text.py
@@ -1118,1014 +1118,52 @@ def normalize_temporal_constructor_text(text: str) -> Optional[str]:
     return None
 
 
-@dataclass(frozen=True)
-class _TemporalValue:
-    kind: str
-    date_value: Optional[py_date]
-    hour: int = 0
-    minute: int = 0
-    second: int = 0
-    nanosecond: int = 0
-    tz_suffix: Optional[str] = None
-
-
-@dataclass(frozen=True)
-class _WideTemporalValue:
-    kind: str
-    year: int
-    month: int
-    day: int
-    hour: int = 0
-    minute: int = 0
-    second: int = 0
-    nanosecond: int = 0
-
-
-_ZONE_SUFFIX_RE = re.compile(r"^(?P<core>.+?)(?:\[(?P<zone>[^\]]+)\])?$")
-_DURATION_TOKEN_RE = re.compile(r"([+-]?\d+(?:\.\d+)?)([YMWDHMS])")
-_DATE_TRUNCATION_UNITS = frozenset({"millennium", "century", "decade", "year", "weekYear", "quarter", "month", "week", "day"})
-_DAY_TIME_DURATION_TOKEN_RE = re.compile(r"([+-]?\d+(?:\.\d+)?)([WDHMS])")
-
-
-def _format_localtime_parts(hour: int, minute: int, second: int, nanosecond: int) -> str:
-    out = f"{hour:02d}:{minute:02d}"
-    if second != 0 or nanosecond != 0:
-        out += f":{second:02d}{_normalize_fraction(nanosecond)}"
-    return out
-
-
-def _format_localdatetime_parts(
-    value_date: py_date,
-    hour: int,
-    minute: int,
-    second: int,
-    nanosecond: int,
-) -> str:
-    return f"{_format_date(value_date.year, value_date.month, value_date.day)}T{_format_localtime_parts(hour, minute, second, nanosecond)}"
-
-
-def parse_temporal_sort_duration_components(text: str) -> Optional[tuple[int, int]]:
-    stripped = text.strip()
-    prefix_sign = 1
-    if stripped.startswith("-P"):
-        prefix_sign = -1
-        body = stripped[2:]
-    elif stripped.startswith("P"):
-        body = stripped[1:]
-    else:
-        return None
-    if body == "":
-        return None
-
-    if "T" in body:
-        date_part, time_part = body.split("T", 1)
-    else:
-        date_part, time_part = body, ""
-
-    month_shift = 0
-    total = Decimal(0)
-
-    def _consume(part: str, allowed_units: set[str]) -> Optional[list[tuple[Decimal, str]]]:
-        if part == "":
-            return []
-        pos = 0
-        out: list[tuple[Decimal, str]] = []
-        for match in _DAY_TIME_DURATION_TOKEN_RE.finditer(part):
-            if match.start() != pos:
-                return None
-            value_txt, unit = match.groups()
-            if unit not in allowed_units:
-                return None
-            out.append((Decimal(value_txt), unit))
-            pos = match.end()
-        if pos != len(part):
-            return None
-        return out
-
-    date_tokens = _consume(date_part, {"Y", "M", "W", "D"})
-    time_tokens = _consume(time_part, {"H", "M", "S"})
-    if date_tokens is None or time_tokens is None:
-        return None
-
-    for value, unit in date_tokens:
-        if unit == "Y":
-            if value != int(value):
-                return None
-            month_shift += int(value) * 12
-        elif unit == "M":
-            if value != int(value):
-                return None
-            month_shift += int(value)
-        elif unit == "W":
-            total += value * Decimal(7 * 24 * 60 * 60 * 1_000_000_000)
-        elif unit == "D":
-            total += value * Decimal(24 * 60 * 60 * 1_000_000_000)
-
-    for value, unit in time_tokens:
-        if unit == "H":
-            total += value * Decimal(60 * 60 * 1_000_000_000)
-        elif unit == "M":
-            total += value * Decimal(60 * 1_000_000_000)
-        elif unit == "S":
-            total += value * Decimal(1_000_000_000)
-
-    return month_shift * prefix_sign, int((total * prefix_sign).to_integral_value())
-
-
-def parse_day_time_duration_nanoseconds(text: str) -> Optional[int]:
-    parsed = parse_temporal_sort_duration_components(text)
-    if parsed is None:
-        return None
-    month_shift, nanosecond_shift = parsed
-    if month_shift != 0:
-        return None
-    return nanosecond_shift
-
-
-def _split_zone_name(text: str) -> tuple[str, Optional[str]]:
-    match = _ZONE_SUFFIX_RE.fullmatch(text)
-    if match is None:
-        return text, None
-    return match.group("core"), match.group("zone")
-
-
-def _parse_temporal_value(text: str) -> Optional[_TemporalValue]:
-    stripped = text.strip()
-    if stripped.startswith(("P", "-P")):
-        return None
-    if "T" in stripped:
-        date_text, time_text = stripped.split("T", 1)
-        value_date = _base_date_from_text(date_text)
-        if value_date is None:
-            return None
-        time_core, zone_name = _split_zone_name(time_text)
-        parts = _base_time_parts_from_text(time_core)
-        if parts is None:
-            return None
-        tz = cast(Optional[str], parts.get("tz"))
-        tz_suffix = tz if zone_name is None else ("" if tz is None else tz) + f"[{zone_name}]"
-        return _TemporalValue(
-            kind="datetime" if tz_suffix is not None else "localdatetime",
-            date_value=value_date,
-            hour=cast(int, parts["hour"]),
-            minute=cast(int, parts["minute"]),
-            second=cast(int, parts["second"]),
-            nanosecond=cast(int, parts["nanosecond"]),
-            tz_suffix=tz_suffix,
-        )
-    if ":" in stripped:
-        time_core, zone_name = _split_zone_name(stripped)
-        parts = _base_time_parts_from_text(time_core)
-        if parts is None:
-            return None
-        tz = cast(Optional[str], parts.get("tz"))
-        tz_suffix = tz if zone_name is None else ("" if tz is None else tz) + f"[{zone_name}]"
-        return _TemporalValue(
-            kind="time" if tz_suffix is not None else "localtime",
-            date_value=None,
-            hour=cast(int, parts["hour"]),
-            minute=cast(int, parts["minute"]),
-            second=cast(int, parts["second"]),
-            nanosecond=cast(int, parts["nanosecond"]),
-            tz_suffix=tz_suffix,
-        )
-    value_date = _base_date_from_text(stripped)
-    if value_date is None:
-        return None
-    return _TemporalValue(kind="date", date_value=value_date)
-
-
-def _truncate_year(year: int, span: int) -> int:
-    return (year // span) * span
-
-
-def _is_leap_year(year: int) -> bool:
-    return (year % 4 == 0) and ((year % 100 != 0) or (year % 400 == 0))
-
-
-def _days_in_month(year: int, month: int) -> int:
-    if month in {1, 3, 5, 7, 8, 10, 12}:
-        return 31
-    if month in {4, 6, 9, 11}:
-        return 30
-    if month == 2:
-        return 29 if _is_leap_year(year) else 28
-    raise ValueError(f"invalid month: {month}")
-
-
-def _days_from_civil(year: int, month: int, day: int) -> int:
-    adjusted_year = year - (1 if month <= 2 else 0)
-    era = adjusted_year // 400
-    year_of_era = adjusted_year - (era * 400)
-    month_prime = month + (-3 if month > 2 else 9)
-    day_of_year = ((153 * month_prime) + 2) // 5 + day - 1
-    day_of_era = year_of_era * 365 + year_of_era // 4 - year_of_era // 100 + day_of_year
-    return era * 146097 + day_of_era - 719468
-
-
-def _parse_wide_temporal_value(text: str) -> Optional[_WideTemporalValue]:
-    stripped = text.strip()
-    localdatetime_match = _WIDE_LOCALDATETIME_TEXT_RE.fullmatch(stripped)
-    if localdatetime_match is not None:
-        year = int(localdatetime_match.group("year"))
-        month = int(localdatetime_match.group("month"))
-        day = int(localdatetime_match.group("day"))
-        hour = int(localdatetime_match.group("hour"))
-        minute = int(localdatetime_match.group("minute"))
-        second = int(localdatetime_match.group("second") or "0")
-        if not 1 <= month <= 12 or not 1 <= day <= _days_in_month(year, month):
-            return None
-        if hour > 23 or minute > 59 or second > 59:
-            return None
-        return _WideTemporalValue(
-            kind="localdatetime",
-            year=year,
-            month=month,
-            day=day,
-            hour=hour,
-            minute=minute,
-            second=second,
-            nanosecond=_parse_fraction_to_nanos(localdatetime_match.group("frac")),
-        )
-    date_match = _WIDE_DATE_TEXT_RE.fullmatch(stripped)
-    if date_match is None:
-        return None
-    year = int(date_match.group("year"))
-    month = int(date_match.group("month"))
-    day = int(date_match.group("day"))
-    if not 1 <= month <= 12 or not 1 <= day <= _days_in_month(year, month):
-        return None
-    return _WideTemporalValue(kind="date", year=year, month=month, day=day)
-
-
-def _format_large_time_only_duration(total_nanoseconds: int) -> str:
-    if total_nanoseconds == 0:
-        return "PT0S"
-    sign = -1 if total_nanoseconds < 0 else 1
-    remaining = abs(total_nanoseconds)
-    hours, remaining = divmod(remaining, 3_600_000_000_000)
-    minutes, remaining = divmod(remaining, 60_000_000_000)
-    seconds, nanoseconds = divmod(remaining, 1_000_000_000)
-
-    def _signed(value: int) -> str:
-        return f"{'-' if sign < 0 else ''}{value}"
-
-    parts = ["PT"]
-    if hours:
-        parts.append(f"{_signed(hours)}H")
-    if minutes:
-        parts.append(f"{_signed(minutes)}M")
-    if seconds or nanoseconds or len(parts) == 1:
-        if nanoseconds:
-            frac = str(nanoseconds).rjust(9, "0").rstrip("0")
-            parts.append(f"{_signed(seconds)}.{frac}S")
-        else:
-            parts.append(f"{_signed(seconds)}S")
-    return "".join(parts)
-
-
-def _fold_large_year_duration_function_call(
-    fn_name: str,
-    start_text: str,
-    end_text: str,
-) -> Optional[str]:
-    start_value = _parse_wide_temporal_value(start_text)
-    end_value = _parse_wide_temporal_value(end_text)
-    if start_value is None or end_value is None:
-        return None
-    if fn_name == "duration.between":
-        if start_value.kind != "date" or end_value.kind != "date":
-            return None
-        if (end_value.year, end_value.month, end_value.day) < (start_value.year, start_value.month, start_value.day):
-            return None
-        years = end_value.year - start_value.year
-        months = end_value.month - start_value.month
-        days = end_value.day - start_value.day
-        if days < 0:
-            months -= 1
-            prev_year = end_value.year
-            prev_month = end_value.month - 1
-            if prev_month == 0:
-                prev_year -= 1
-                prev_month = 12
-            days += _days_in_month(prev_year, prev_month)
-        if months < 0:
-            years -= 1
-            months += 12
-        return _format_duration_components(years=years, months=months, days=days)
-    if fn_name == "duration.inseconds":
-        if start_value.kind != "localdatetime" or end_value.kind != "localdatetime":
-            return None
-        start_days = _days_from_civil(start_value.year, start_value.month, start_value.day)
-        end_days = _days_from_civil(end_value.year, end_value.month, end_value.day)
-        start_total = (
-            start_days * 86_400_000_000_000
-            + start_value.hour * 3_600_000_000_000
-            + start_value.minute * 60_000_000_000
-            + start_value.second * 1_000_000_000
-            + start_value.nanosecond
-        )
-        end_total = (
-            end_days * 86_400_000_000_000
-            + end_value.hour * 3_600_000_000_000
-            + end_value.minute * 60_000_000_000
-            + end_value.second * 1_000_000_000
-            + end_value.nanosecond
-        )
-        return _format_large_time_only_duration(end_total - start_total)
-    return None
-
-
-def _truncate_date_value(source_date: py_date, unit: str, overrides: dict[str, str]) -> Optional[py_date]:
-    if unit == "millennium":
-        base = py_date(_truncate_year(source_date.year, 1000), 1, 1)
-    elif unit == "century":
-        base = py_date(_truncate_year(source_date.year, 100), 1, 1)
-    elif unit == "decade":
-        base = py_date(_truncate_year(source_date.year, 10), 1, 1)
-    elif unit == "year":
-        base = py_date(source_date.year, 1, 1)
-    elif unit == "quarter":
-        base = py_date(source_date.year, ((source_date.month - 1) // 3) * 3 + 1, 1)
-    elif unit == "month":
-        base = py_date(source_date.year, source_date.month, 1)
-    elif unit == "week":
-        base = source_date - timedelta(days=source_date.isoweekday() - 1)
-        if "dayOfWeek" in overrides:
-            return base + timedelta(days=int(overrides["dayOfWeek"]) - 1)
-        if "day" in overrides:
-            return base + timedelta(days=int(overrides["day"]) - 1)
-    elif unit == "weekYear":
-        base = py_date.fromisocalendar(source_date.isocalendar()[0], 1, 1)
-        if "day" in overrides:
-            return py_date(base.year, 1, 1) + timedelta(days=int(overrides["day"]) - 1)
-        if "dayOfWeek" in overrides:
-            return base + timedelta(days=int(overrides["dayOfWeek"]) - 1)
-    elif unit == "day":
-        base = source_date
-    else:
-        return None
-    fields = {"date": _format_date(base.year, base.month, base.day)}
-    fields.update(overrides)
-    return _date_from_fields(fields)
-
-
-def _truncate_time_parts(
-    source_value: _TemporalValue,
-    unit: str,
-    overrides: dict[str, str],
-) -> tuple[int, int, int, int]:
-    if unit in _DATE_TRUNCATION_UNITS:
-        hour = 0
-        minute = 0
-        second = 0
-        nanosecond = 0
-    else:
-        hour = source_value.hour
-        minute = source_value.minute
-        second = source_value.second
-        nanosecond = source_value.nanosecond
-
-    if unit == "day":
-        hour = 0
-        minute = 0
-        second = 0
-        nanosecond = 0
-    elif unit == "hour":
-        minute = 0
-        second = 0
-        nanosecond = 0
-    elif unit == "minute":
-        second = 0
-        nanosecond = 0
-    elif unit == "second":
-        nanosecond = 0
-    elif unit == "millisecond":
-        nanosecond = (nanosecond // 1_000_000) * 1_000_000
-    elif unit == "microsecond":
-        nanosecond = (nanosecond // 1_000) * 1_000
-
-    if "hour" in overrides:
-        hour = int(overrides["hour"])
-    if "minute" in overrides:
-        minute = int(overrides["minute"])
-    if "second" in overrides:
-        second = int(overrides["second"])
-    if any(key in overrides for key in ("nanosecond", "microsecond", "millisecond")):
-        truncated_millisecond = nanosecond // 1_000_000
-        truncated_microsecond = (nanosecond // 1_000) % 1_000
-        truncated_nanosecond = nanosecond % 1_000
-        millisecond = _parse_int(
-            overrides.get("millisecond", overrides.get("milliseconds")),
-            truncated_millisecond,
-        )
-        microsecond = _parse_int(
-            overrides.get("microsecond", overrides.get("microseconds")),
-            truncated_microsecond,
-        )
-        sub_nanosecond = _parse_int(
-            overrides.get("nanosecond", overrides.get("nanoseconds")),
-            truncated_nanosecond,
-        )
-        if millisecond is None or microsecond is None or sub_nanosecond is None:
-            return hour, minute, second, nanosecond
-        nanosecond = (millisecond * 1_000_000) + (microsecond * 1_000) + sub_nanosecond
-
-    return hour, minute, second, nanosecond
-
-
-def _zone_compatible_local_datetime_text(date_value: py_date, local_time_text: str) -> str:
-    local_dt_text = f"{_format_date(date_value.year, date_value.month, date_value.day)}T{local_time_text}"
-    match = re.fullmatch(r"(?P<prefix>.+?\.\d{6})\d+(?P<suffix>)", local_dt_text)
-    if match is not None:
-        return match.group("prefix")
-    return local_dt_text
-
-
-def _target_timezone_suffix(
-    target_kind: str,
-    source_value: _TemporalValue,
-    overrides: dict[str, str],
-    *,
-    target_date: Optional[py_date],
-    local_time_text: str,
-) -> Optional[str]:
-    if target_kind not in {"time", "datetime"}:
-        return None
-    timezone_text = overrides.get("timezone")
-    if timezone_text is None:
-        return source_value.tz_suffix or "Z"
-    zone_name = _parse_quoted(timezone_text)
-    if zone_name is None and not re.fullmatch(r"Z|[+-]\d{2}(?::?\d{2})?(?::?\d{2})?", timezone_text):
-        zone_name = timezone_text
-    if zone_name is None:
-        return _normalize_offset_text(timezone_text)
-    zone_base_date = target_date or source_value.date_value or py_date(1970, 1, 1)
-    suffix = _zone_suffix(
-        zone_name,
-        _zone_compatible_local_datetime_text(zone_base_date, local_time_text),
-    )
-    return suffix
-
-
-def _fold_temporal_truncate_call(
-    fn_name: str,
-    args: tuple[ExprNode, ...],
-) -> Optional[Literal]:
-    if len(args) != 3:
-        return None
-    unit_node, value_node, override_node = args
-    if not isinstance(unit_node, Literal) or not isinstance(unit_node.value, str):
-        return None
-    if not isinstance(value_node, Literal) or not isinstance(value_node.value, str):
-        return None
-    if not isinstance(override_node, MapLiteral):
-        return None
-
-    overrides: dict[str, str] = {}
-    for key, value in override_node.items:
-        rendered = _render_temporal_arg(value)
-        if rendered is None:
-            return None
-        parsed = _parse_quoted(rendered)
-        overrides[key] = rendered if parsed is None else parsed
-
-    source_value = _parse_temporal_value(value_node.value)
-    if source_value is None:
-        return None
-
-    unit = unit_node.value
-    target_kind = fn_name.split(".", 1)[0]
-    target_date: Optional[py_date] = source_value.date_value
-    if target_kind == "date":
-        if target_date is None:
-            target_date = py_date(1970, 1, 1)
-        target_date = _truncate_date_value(target_date, unit, overrides)
-        if target_date is None:
-            return None
-    elif target_kind in {"datetime", "localdatetime"} and unit in _DATE_TRUNCATION_UNITS:
-        if target_date is None:
-            target_date = py_date(1970, 1, 1)
-        target_date = _truncate_date_value(target_date, unit, overrides)
-        if target_date is None:
-            return None
-
-    hour, minute, second, nanosecond = _truncate_time_parts(source_value, unit, overrides)
-
-    if target_kind == "date":
-        assert target_date is not None
-        return Literal(_format_date(target_date.year, target_date.month, target_date.day))
-
-    local_time_text = _format_localtime_parts(hour, minute, second, nanosecond)
-    if target_kind == "localtime":
-        return Literal(local_time_text)
-
-    tz_suffix = _target_timezone_suffix(
-        target_kind,
-        source_value,
-        overrides,
-        target_date=target_date,
-        local_time_text=local_time_text,
-    )
-
-    if target_kind == "time":
-        return Literal(local_time_text + cast(str, tz_suffix))
-
-    assert target_date is not None
-    local_dt_text = _format_localdatetime_parts(target_date, hour, minute, second, nanosecond)
-    if target_kind == "localdatetime":
-        return Literal(local_dt_text)
-    return Literal(local_dt_text + cast(str, tz_suffix))
-
-
-def _comparable_datetime(
-    value: _TemporalValue,
-    *,
-    include_date: bool,
-    keep_timezone: bool,
-    anchor_date: Optional[py_date] = None,
-    anchor_tz_suffix: Optional[str] = None,
-) -> py_datetime:
-    if include_date:
-        value_date = value.date_value or anchor_date or py_date(1970, 1, 1)
-    else:
-        value_date = anchor_date or py_date(1970, 1, 1)
-    effective_tz_suffix = value.tz_suffix or anchor_tz_suffix
-    naive = py_datetime(
-        value_date.year,
-        value_date.month,
-        value_date.day,
-        value.hour,
-        value.minute,
-        value.second,
-        value.nanosecond // 1_000,
-    )
-    if keep_timezone and effective_tz_suffix is not None:
-        offset = effective_tz_suffix.split("[", 1)[0]
-        zone_match = re.search(r"\[(?P<zone>[^\]]+)\]$", effective_tz_suffix)
-        if zone_match is not None:
-            zone_name = zone_match.group("zone")
-            historical_override = _neo4j_historical_zone_offset(zone_name, naive)
-            if historical_override is not None:
-                return naive.replace(tzinfo=py_timezone(historical_override))
-            zone = _named_timezone(zone_name)
-            if zone is not None:
-                return naive.replace(tzinfo=zone)
-        if offset == "Z":
-            return naive.replace(tzinfo=py_timezone.utc)
-        offset_delta = py_timedelta_from_offset(offset)
-        if offset_delta is not None:
-            return naive.replace(tzinfo=py_timezone(offset_delta))
-        return naive
-    return naive
-
-
-def py_timedelta_from_offset(offset: str) -> Optional[timedelta]:
-    match = re.fullmatch(r"(?P<sign>[+-])(?P<hour>\d{2}):(?P<minute>\d{2})(?::(?P<second>\d{2}))?", offset)
-    if match is None:
-        return None
-    sign = 1 if match.group("sign") == "+" else -1
-    delta = timedelta(
-        hours=int(match.group("hour")),
-        minutes=int(match.group("minute")),
-        seconds=int(match.group("second") or "0"),
-    )
-    return delta if sign > 0 else -delta
-
-
-def _timedelta_total_microseconds(delta: timedelta) -> int:
-    return ((delta.days * 24 * 60 * 60) + delta.seconds) * 1_000_000 + delta.microseconds
-
-
-def _absolute_temporal_delta(start_dt: py_datetime, end_dt: py_datetime) -> timedelta:
-    if start_dt.tzinfo is not None and end_dt.tzinfo is not None:
-        return end_dt.astimezone(py_timezone.utc) - start_dt.astimezone(py_timezone.utc)
-    return end_dt - start_dt
-
-
-def _format_time_only_duration(delta: timedelta) -> str:
-    total_microseconds = _timedelta_total_microseconds(delta)
-    if total_microseconds == 0:
-        return "PT0S"
-    sign = -1 if total_microseconds < 0 else 1
-    remaining = abs(total_microseconds)
-    hours, remaining = divmod(remaining, 3_600_000_000)
-    minutes, remaining = divmod(remaining, 60_000_000)
-    seconds, microseconds = divmod(remaining, 1_000_000)
-
-    def _signed(value: int) -> str:
-        return f"{'-' if sign < 0 else ''}{value}"
-
-    parts = ["PT"]
-    if hours:
-        parts.append(f"{_signed(hours)}H")
-    if minutes:
-        parts.append(f"{_signed(minutes)}M")
-    if seconds or microseconds or len(parts) == 1:
-        if microseconds:
-            frac = str(microseconds).rjust(6, "0").rstrip("0")
-            parts.append(f"{_signed(seconds)}.{frac}S")
-        else:
-            parts.append(f"{_signed(seconds)}S")
-    return "".join(parts)
-
-
-def _format_duration_components(
-    *,
-    years: int = 0,
-    months: int = 0,
-    days: int = 0,
-    hours: int = 0,
-    minutes: int = 0,
-    seconds: int = 0,
-    microseconds: int = 0,
-) -> str:
-    parts = ["P"]
-    if years:
-        parts.append(f"{years}Y")
-    if months:
-        parts.append(f"{months}M")
-    if days:
-        parts.append(f"{days}D")
-    time_parts: list[str] = []
-    if hours:
-        time_parts.append(f"{hours}H")
-    if minutes:
-        time_parts.append(f"{minutes}M")
-    if seconds or microseconds:
-        if microseconds:
-            frac = str(abs(microseconds)).rjust(6, "0").rstrip("0")
-            time_parts.append(f"{seconds}.{frac}S")
-        else:
-            time_parts.append(f"{seconds}S")
-    if time_parts:
-        parts.append("T")
-        parts.extend(time_parts)
-    if parts == ["P"]:
-        return "PT0S"
-    return "".join(parts)
-
-
-def _fold_duration_between(
-    start_value: _TemporalValue,
-    end_value: _TemporalValue,
-) -> str:
-    include_date = start_value.date_value is not None and end_value.date_value is not None
-    keep_timezone = start_value.tz_suffix is not None and end_value.tz_suffix is not None
-    anchor_date = start_value.date_value or end_value.date_value
-    anchor_tz_suffix = start_value.tz_suffix or end_value.tz_suffix
-    effective_keep_timezone = keep_timezone or anchor_tz_suffix is not None
-    start_dt = _comparable_datetime(
-        start_value,
-        include_date=include_date,
-        keep_timezone=effective_keep_timezone,
-        anchor_date=anchor_date,
-        anchor_tz_suffix=anchor_tz_suffix if start_value.tz_suffix is None else None,
-    )
-    end_dt = _comparable_datetime(
-        end_value,
-        include_date=include_date,
-        keep_timezone=effective_keep_timezone,
-        anchor_date=anchor_date,
-        anchor_tz_suffix=anchor_tz_suffix if end_value.tz_suffix is None else None,
-    )
-    delta = _absolute_temporal_delta(start_dt, end_dt)
-    if not include_date:
-        return _format_time_only_duration(delta)
-
-    from dateutil.relativedelta import relativedelta  # type: ignore[import]
-
-    rel = relativedelta(end_dt, start_dt)
-    if rel.years == 0 and rel.months == 0:
-        return _format_signed_day_time_duration(_timedelta_total_microseconds(delta) * 1_000)
-    return _format_duration_components(
-        years=rel.years,
-        months=rel.months,
-        days=rel.days,
-        hours=rel.hours,
-        minutes=rel.minutes,
-        seconds=rel.seconds,
-        microseconds=rel.microseconds,
-    )
-
-
-def _fold_duration_in_months(
-    start_value: _TemporalValue,
-    end_value: _TemporalValue,
-) -> str:
-    if start_value.date_value is None or end_value.date_value is None:
-        return "PT0S"
-    keep_timezone = start_value.tz_suffix is not None and end_value.tz_suffix is not None
-    start_dt = _comparable_datetime(start_value, include_date=True, keep_timezone=keep_timezone)
-    end_dt = _comparable_datetime(end_value, include_date=True, keep_timezone=keep_timezone)
-    from dateutil.relativedelta import relativedelta  # type: ignore[import]
-
-    rel = relativedelta(end_dt, start_dt)
-    return _format_duration_components(years=rel.years, months=rel.months)
-
-
-def _fold_duration_in_days(
-    start_value: _TemporalValue,
-    end_value: _TemporalValue,
-) -> str:
-    if start_value.date_value is None or end_value.date_value is None:
-        return "PT0S"
-    keep_timezone = start_value.tz_suffix is not None and end_value.tz_suffix is not None
-    start_dt = _comparable_datetime(start_value, include_date=True, keep_timezone=keep_timezone)
-    end_dt = _comparable_datetime(end_value, include_date=True, keep_timezone=keep_timezone)
-    total_microseconds = _timedelta_total_microseconds(_absolute_temporal_delta(start_dt, end_dt))
-    days = int(total_microseconds / (24 * 60 * 60 * 1_000_000))
-    return _format_duration_components(days=days)
-
-
-def _fold_duration_in_seconds(
-    start_value: _TemporalValue,
-    end_value: _TemporalValue,
-) -> str:
-    include_date = start_value.date_value is not None and end_value.date_value is not None
-    keep_timezone = start_value.tz_suffix is not None and end_value.tz_suffix is not None
-    anchor_date = start_value.date_value or end_value.date_value
-    anchor_tz_suffix = start_value.tz_suffix or end_value.tz_suffix
-    effective_keep_timezone = keep_timezone or anchor_tz_suffix is not None
-    start_dt = _comparable_datetime(
-        start_value,
-        include_date=include_date,
-        keep_timezone=effective_keep_timezone,
-        anchor_date=anchor_date,
-        anchor_tz_suffix=anchor_tz_suffix if start_value.tz_suffix is None else None,
-    )
-    end_dt = _comparable_datetime(
-        end_value,
-        include_date=include_date,
-        keep_timezone=effective_keep_timezone,
-        anchor_date=anchor_date,
-        anchor_tz_suffix=anchor_tz_suffix if end_value.tz_suffix is None else None,
-    )
-    return _format_time_only_duration(_absolute_temporal_delta(start_dt, end_dt))
-
-
-def _fold_duration_function_call(
-    fn_name: str,
-    args: tuple[ExprNode, ...],
-) -> Optional[Literal]:
-    if len(args) != 2:
-        return None
-    if any(isinstance(arg, Literal) and arg.value is None for arg in args):
-        return Literal(None)
-    if not all(isinstance(arg, Literal) and isinstance(arg.value, str) for arg in args):
-        return None
-    start_text = cast(str, cast(Literal, args[0]).value)
-    end_text = cast(str, cast(Literal, args[1]).value)
-    try:
-        start_value = _parse_temporal_value(start_text)
-        end_value = _parse_temporal_value(end_text)
-    except ValueError:
-        start_value = None
-        end_value = None
-    if start_value is None or end_value is None:
-        large_year = _fold_large_year_duration_function_call(fn_name, start_text, end_text)
-        if large_year is not None:
-            return Literal(large_year)
-        return None
-    if fn_name == "duration.between":
-        return Literal(_fold_duration_between(start_value, end_value))
-    if fn_name == "duration.inmonths":
-        return Literal(_fold_duration_in_months(start_value, end_value))
-    if fn_name == "duration.indays":
-        return Literal(_fold_duration_in_days(start_value, end_value))
-    if fn_name == "duration.inseconds":
-        return Literal(_fold_duration_in_seconds(start_value, end_value))
-    return None
-
-
-def _fold_datetime_epoch_function_call(
-    fn_name: str,
-    args: tuple[ExprNode, ...],
-) -> Optional[Literal]:
-    if fn_name not in {"datetime.fromepoch", "datetime.fromepochmillis"}:
-        return None
-    if any(isinstance(arg, Literal) and arg.value is None for arg in args):
-        return Literal(None)
-    if not all(isinstance(arg, Literal) and isinstance(arg.value, int) and not isinstance(arg.value, bool) for arg in args):
-        return None
-
-    epoch = py_datetime(1970, 1, 1)
-    if fn_name == "datetime.fromepochmillis":
-        if len(args) != 1:
-            return None
-        total_nanoseconds = cast(int, cast(Literal, args[0]).value) * 1_000_000
-    else:
-        if len(args) not in {1, 2}:
-            return None
-        seconds_value = cast(int, cast(Literal, args[0]).value)
-        nanoseconds_value = cast(int, cast(Literal, args[1]).value) if len(args) == 2 else 0
-        total_nanoseconds = (seconds_value * 1_000_000_000) + nanoseconds_value
-
-    seconds_part, nanoseconds_part = divmod(total_nanoseconds, 1_000_000_000)
-    dt = epoch + timedelta(seconds=seconds_part, microseconds=nanoseconds_part // 1_000)
-    rendered = _format_localdatetime_parts(
-        dt.date(),
-        dt.hour,
-        dt.minute,
-        dt.second,
-        int(nanoseconds_part),
-    )
-    return Literal(rendered + "Z")
-
-
-def resolve_duration_text_property(duration_text: str, prop: str) -> Optional[str]:
-    if not duration_text.startswith(("P", "-P")):
-        return None
-    days_value = 0
-    time_ns = 0
-    body = duration_text[1:] if duration_text.startswith("P") else duration_text[2:]
-    date_part, _, time_part = body.partition("T")
-    for number_text, unit in _DURATION_TOKEN_RE.findall(date_part):
-        if unit == "D":
-            days_value = int(Decimal(number_text))
-    for number_text, unit in _DURATION_TOKEN_RE.findall(time_part):
-        if unit == "H":
-            time_ns += int(Decimal(number_text) * Decimal(3_600_000_000_000))
-        elif unit == "M":
-            time_ns += int(Decimal(number_text) * Decimal(60_000_000_000))
-        elif unit == "S":
-            time_ns += int(Decimal(number_text) * Decimal(1_000_000_000))
-    if prop == "days":
-        return str(days_value)
-    if prop == "seconds":
-        return str(time_ns // 1_000_000_000)
-    if prop == "nanosecondsOfSecond":
-        seconds_value = time_ns // 1_000_000_000
-        return str(time_ns - (seconds_value * 1_000_000_000))
-    return None
-
-
-def rewrite_temporal_constructors_in_expr(expr_text: str) -> str:
-    current_dt = py_datetime.now().astimezone()
-
-    def _replace_current(match: re.Match[str]) -> str:
-        normalized = _current_temporal_literal(match.group("fn"), current_dt)
-        if normalized is None:
-            return match.group(0)
-        escaped = normalized.replace("\\", "\\\\").replace("'", "\\'")
-        return f"'{escaped}'"
-
-    def _replace(match: re.Match[str]) -> str:
-        normalized = normalize_temporal_constructor_text(match.group(0))
-        if normalized is None:
-            return match.group(0)
-        escaped = normalized.replace("\\", "\\\\").replace("'", "\\'")
-        return f"'{escaped}'"
-
-    rewritten = CURRENT_TEMPORAL_CALL_EXPR_RE.sub(_replace_current, expr_text)
-    return TEMPORAL_CALL_EXPR_RE.sub(_replace, rewritten)
-
-
-def _render_temporal_arg(node: ExprNode) -> Optional[str]:
-    if isinstance(node, Literal):
-        value = node.value
-        if value is None:
-            return "null"
-        if isinstance(value, bool):
-            return "true" if value else "false"
-        if isinstance(value, (int, float)) and not isinstance(value, bool):
-            return str(value)
-        if isinstance(value, str):
-            escaped = value.replace("\\", "\\\\").replace("'", "\\'")
-            return f"'{escaped}'"
-        return None
-    if isinstance(node, MapLiteral):
-        parts: list[str] = []
-        for key, value in node.items:
-            rendered = _render_temporal_arg(value)
-            if rendered is None:
-                return None
-            parts.append(f"{key}: {rendered}")
-        return "{" + ", ".join(parts) + "}"
-    if isinstance(node, ListLiteral):
-        rendered_items: list[str] = []
-        for item in node.items:
-            rendered = _render_temporal_arg(item)
-            if rendered is None:
-                return None
-            rendered_items.append(rendered)
-        return "[" + ", ".join(rendered_items) + "]"
-    return None
-
-
-def fold_temporal_constructor_ast(node: ExprNode) -> ExprNode:
-    current_dt = py_datetime.now().astimezone()
-
-    def _fold(inner: ExprNode) -> ExprNode:
-        if isinstance(inner, (Identifier, Literal, Wildcard)):
-            return inner
-        if isinstance(inner, UnaryOp):
-            return UnaryOp(inner.op, _fold(inner.operand))
-        if isinstance(inner, BinaryOp):
-            return BinaryOp(
-                inner.op,
-                _fold(inner.left),
-                _fold(inner.right),
-            )
-        if isinstance(inner, IsNullOp):
-            return IsNullOp(_fold(inner.value), negated=inner.negated)
-        if isinstance(inner, FunctionCall):
-            args = tuple(_fold(arg) for arg in inner.args)
-            rewritten = FunctionCall(inner.name, args, distinct=inner.distinct)
-            if not inner.distinct and len(args) == 0 and inner.name in {
-                "date",
-                "localtime",
-                "time",
-                "localdatetime",
-                "datetime",
-            }:
-                current_literal = _current_temporal_literal(inner.name, current_dt)
-                if current_literal is not None:
-                    return Literal(current_literal)
-            if not inner.distinct and inner.name == "tostring" and len(args) == 1 and isinstance(args[0], Literal):
-                value = args[0].value
-                if value is None:
-                    return Literal(None)
-                if isinstance(value, bool):
-                    return Literal("true" if value else "false")
-                return Literal(str(value))
-            if not inner.distinct and len(args) == 1 and inner.name in {
-                "date",
-                "localtime",
-                "time",
-                "localdatetime",
-                "datetime",
-                "duration",
-            } and isinstance(args[0], Literal) and args[0].value is None:
-                return Literal(None)
-            if not inner.distinct and len(args) == 1 and inner.name in {"date", "localtime", "time", "localdatetime", "datetime", "duration"}:
-                rendered_arg = _render_temporal_arg(args[0])
-                if rendered_arg is not None:
-                    normalized = normalize_temporal_constructor_text(f"{inner.name}({rendered_arg})")
-                    if normalized is not None:
-                        return Literal(normalized)
-            if not inner.distinct and inner.name in {
-                "date.truncate",
-                "localtime.truncate",
-                "time.truncate",
-                "localdatetime.truncate",
-                "datetime.truncate",
-            }:
-                folded = _fold_temporal_truncate_call(inner.name, args)
-                if folded is not None:
-                    return folded
-            if not inner.distinct and inner.name in {
-                "duration.between",
-                "duration.inmonths",
-                "duration.indays",
-                "duration.inseconds",
-            }:
-                folded = _fold_duration_function_call(inner.name, args)
-                if folded is not None:
-                    return folded
-            if not inner.distinct and inner.name in {
-                "datetime.fromepoch",
-                "datetime.fromepochmillis",
-            }:
-                folded = _fold_datetime_epoch_function_call(inner.name, args)
-                if folded is not None:
-                    return folded
-            return rewritten
-        if isinstance(inner, CaseWhen):
-            return CaseWhen(
-                _fold(inner.condition),
-                _fold(inner.when_true),
-                _fold(inner.when_false),
-            )
-        if isinstance(inner, QuantifierExpr):
-            return QuantifierExpr(
-                inner.fn,
-                inner.var,
-                _fold(inner.source),
-                _fold(inner.predicate),
-            )
-        if isinstance(inner, ListComprehension):
-            return ListComprehension(
-                inner.var,
-                _fold(inner.source),
-                predicate=None if inner.predicate is None else _fold(inner.predicate),
-                projection=None if inner.projection is None else _fold(inner.projection),
-            )
-        if isinstance(inner, ListLiteral):
-            return ListLiteral(tuple(_fold(item) for item in inner.items))
-        if isinstance(inner, MapLiteral):
-            return MapLiteral(tuple((key, _fold(value)) for key, value in inner.items))
-        if isinstance(inner, SubscriptExpr):
-            return SubscriptExpr(
-                _fold(inner.value),
-                _fold(inner.key),
-            )
-        if isinstance(inner, SliceExpr):
-            return SliceExpr(
-                _fold(inner.value),
-                None if inner.start is None else _fold(inner.start),
-                None if inner.stop is None else _fold(inner.stop),
-            )
-        return inner
-
-    return _fold(node)
+# Compatibility facade: keep the historical temporal_text import path stable
+# while domain-specific implementation lives under graphistry.compute.gfql.temporal.
+from graphistry.compute.gfql.temporal.values import (  # noqa: E402
+    _TemporalValue,
+    _WideTemporalValue,
+    _absolute_temporal_delta,
+    _comparable_datetime,
+    _days_from_civil,
+    _days_in_month,
+    _format_localdatetime_parts,
+    _format_localtime_parts,
+    _is_leap_year,
+    _parse_temporal_value,
+    _parse_wide_temporal_value,
+    _split_zone_name,
+    _truncate_year,
+    py_timedelta_from_offset,
+)
+from graphistry.compute.gfql.temporal.durations import (  # noqa: E402
+    _DAY_TIME_DURATION_TOKEN_RE,
+    _DURATION_TOKEN_RE,
+    _fold_duration_between,
+    _fold_duration_function_call,
+    _fold_duration_in_days,
+    _fold_duration_in_months,
+    _fold_duration_in_seconds,
+    _fold_large_year_duration_function_call,
+    _format_duration_components,
+    _format_large_time_only_duration,
+    _format_time_only_duration,
+    _timedelta_total_microseconds,
+    parse_day_time_duration_nanoseconds,
+    parse_temporal_sort_duration_components,
+    resolve_duration_text_property,
+)
+from graphistry.compute.gfql.temporal.truncation import (  # noqa: E402
+    _DATE_TRUNCATION_UNITS,
+    _fold_temporal_truncate_call,
+    _target_timezone_suffix,
+    _truncate_date_value,
+    _truncate_time_parts,
+    _zone_compatible_local_datetime_text,
+)
+from graphistry.compute.gfql.temporal.rendering import _render_temporal_arg  # noqa: E402
+from graphistry.compute.gfql.temporal.folding import (  # noqa: E402
+    _fold_datetime_epoch_function_call,
+    fold_temporal_constructor_ast,
+    rewrite_temporal_constructors_in_expr,
+)
